This app has the fully integrated alarm, update, warning and setting services.
Alarm-> Two different alarms, one for update and another for warning messages with their corresponding broadcasting methods for backgroud processes.

Settings-> Allows to use the update time. If customer would like warning message as soon as something goes out of range. If notitfications are active it allows to choose a vibrate, sound and light notification options. UI was created and array.xml was created

Current_Data-> BUGS were fixed(No more toast messages everytime app runs in background only while user is on the current activity). Class was modified to allow multithreading.

Refactor-> Current Data was refactored as well as util class. NotificationGeneator had to be refactored due to changes in data retrival and util class. Settings was refactored to allow backgtound proccess to run at user's request not a data update.

TESTS-> Espresso testing was done in another branch as discussed on Sunday July 24 during our weekly meeting. All the additions in this branch use system call and Firebase. There is no way we can test them with junit tests or espresso tests. Solution-> dowload an emulator for JUNIT tests that allows systems calls but this was disccussed with TA and said not to worry. For database testing, we need to use mockito tests but this does not preserve the data base structure nor connectivity. Another solution is to upgrade our GOOGLE membership and use FIREBASE LAB TEST

TODO: INTEGRATION WITH CHAT AND HISTORICAL